output of simulating OT protocol
Opentrons has developed an opentrons_simulate function, which can be used as a debugging tool to check the expected instructions which will be sent to the OT-2 robot. If there are errors in the code, the opentrons_simulate function will raise an Error. The output from running the simulation on the OT-2_ProxyViscometerProtocol.py file is presented here.